+++ typings/global.d.ts
@@ -1,0 +1,24 @@
+declare var process: {
+    env: {
+        PUBLIC_URL: string,
+        NODE_ENV: string
+        APP_ENV: string
+        BASEURL: string
+    }
+}
+
+declare interface PlainObject {
+    [propName: string]: any
+}
+
+declare interface BooleanObject {
+    [propName: string]: boolean
+}
+
+declare interface StringObject {
+    [propName: string]: string
+}
+
+declare interface NumberObject {
+    [propName: string]: number
+}