+++ CHANGELOG.md
@@ -1,4 +1,5 @@
-# 0.1.0 (2019-03-15)
+## 0.1.0 (2019-03-15)
 
-
-
+* chore(project): add docs & changelog ([2452baf](https://github.com/yangfan86/react-ts-demo/commit/2452baf))
+* create-react-app eject first commit ([ec9e642](https://github.com/yangfan86/react-ts-demo/commit/ec9e642))
+* Initial commit from Create React App ([9987108](https://github.com/yangfan86/react-ts-demo/commit/9987108))